Refactor DotRelationship classes + introduced DotRelationship abstract class + introduced UMLRelationships Enumerator for DotRelationship interface
