+++ src/dot_formatter/dot_inheritance.py
@@ -1,11 +1,9 @@
-from dot_formatter import DotAssociation
+from dot_formatter import DotRelationship
 
 
-class DotInheritance(DotAssociation):
+class DotInheritance(DotRelationship):
 
     def _set_options(self, label: str, arrowtail: str = 'empty') -> None:
-        # Override to introduce the arrowtail option, which defaults
-        # to `empty` (arrow outline) for inheritance relationship.
         self.options = {
             'label': label,
             'arrowtail': arrowtail
