+++ src/dot_formatter/dot_relationship.py
@@ -1,0 +1,22 @@
+from abc import ABC, abstractmethod
+
+
+class DotRelationship(ABC):
+
+    def __init__(self, source_class: str, target_class: str, label: str = ''):
+        self.source_class = source_class
+        self.target_class = target_class
+        self._set_options(label)
+
+    @abstractmethod
+    def _set_options(self, label: str) -> None:
+        pass
+
+    def __str__(self) -> str:
+        return ('\n%s -> %s' % (self.target_class, self.source_class)
+                + self._get_options())
+
+    def _get_options(self) -> str:
+        return ' [\n%s\n]' % '\n'.join(
+            ['%s="%s"' % (key, value) for key, value in self.options.items()]
+        )
